requirements.txt
dnspython>=2.0.0
python-dotenv>=0.21.0
requests>=2.28.0